requirements.txt
fastapi==0.95.2
uvicorn==0.20.0
python-dotenv==1.0.0
chromadb==0.3.29
sentence-transformers
requests
